server/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
amqp==5.2.0
annotated-types==0.7.0
anthropic==0.34.2
anyio==4.5.0
asgiref==3.8.1
atomicwrites==1.4.1
attrs==24.3.0
backoff==2.2.1
bcrypt==4.2.0
beautifulsoup4==4.12.3
bidict==0.23.1
billiard==4.2.0
build==1.2.2
cachetools==5.5.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.6
chromadb==0.5.11
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.8.2
cron_descriptor==2.0.6
decorator==4.4.2
defusedxml==0.7.1
Deprecated==1.2.14
distro==1.9.0
dj-database-url==2.2.0
Django==5.1.1
django-celery-beat==2.8.1
django-celery-results==2.5.1
django-cors-headers==4.4.0
django-model-utils==4.3.1
django-redis==5.4.0
django-timezone-field==7.2.1
djangorestframework==3.15.2
durationpy==0.7
elevenlabs==1.50.3
Faker==33.0.0
fastapi==0.115.0
filelock==3.16.1
firecrawl-py==4.6.0
flatbuffers==24.3.25
flower==2.0.1
frozenlist==1.8.0
fsspec==2024.9.0
gevent==24.2.1
google-auth==2.35.0
googleapis-common-protos==1.65.0
greenlet==3.1.1
grpcio==1.66.2
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
huggingface-hub==0.25.0
humanfriendly==10.0
humanize==4.10.0
idna==3.10
imageio==2.35.1
imageio-ffmpeg==0.5.1
importlib_metadata==8.4.0
importlib_resources==6.4.5
iniconfig==2.0.0
Jinja2==3.1.4
jiter==0.13.0
kombu==5.4.2
kubernetes==31.0.0
llvmlite==0.43.0
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mmh3==5.0.1
mock==5.1.0
monotonic==1.6
more-itertools==10.5.0
moviepy==1.0.3
mpmath==1.3.0
multidict==6.7.0
nest-asyncio==1.6.0
networkx==3.3
numba==0.60.0
numpy==2.0.2
oauthlib==3.2.2
onnxruntime==1.19.2
openai==2.20.0
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-asgi==0.48b0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
orjson==3.10.7
overrides==7.7.0
packaging==24.1
pathspec==0.12.1
pillow==10.4.0
pluggy==1.5.0
posthog==3.6.6
proglog==0.1.10
prometheus_client==0.21.0
prompt_toolkit==3.0.47
propcache==0.4.1
protobuf==4.25.5
psutil==6.0.0
psycopg2-binary==2.9.9
py==1.11.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
Pygments==2.18.0
PyMuPDF==1.24.10
PyMuPDFb==1.24.10
pypandoc==1.14
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
pytest==6.2.5
pytest-testdox==3.1.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-engineio==4.9.1
python-multipart==0.0.9
python-socketio==5.11.4
pytz==2024.2
PyYAML==6.0.2
redis==5.0.8
regex==2024.9.11
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.8.1
rsa==4.9
runwayml==2.1.7
serpy==0.3.1
setuptools==75.1.0
shellingham==1.5.4
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.1
starlette==0.38.5
sympy==1.13.3
tenacity==9.0.0
tiktoken==0.7.0
tokenizers==0.20.0
toml==0.10.2
torch==2.4.1
tornado==6.4.1
tqdm==4.66.5
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
uvicorn==0.30.6
vine==5.1.0
watchfiles==0.24.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==13.1
whitenoise==6.7.0
wrapt==1.16.0
wsproto==1.2.0
yarl==1.22.0
youtube-transcript-api==1.2.4
zipp==3.20.2
zope.event==5.0
zope.interface==7.0.3
